+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+- Add mergeSchema option to delta write block
 
 ## [2.7.1] - 2021-09-30
 ### Changed